Update grabbed attributes
The plugin previous would grab the username and add discords discriminator onto the end of it, such as "Username#2422". However, discord did away with discriminators last year, leaving all usernames added as "Username#0" now. I've removed the discriminator because of this and added the additional pull for the id property.